% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cl_test_t.R
\name{Cl_test_t}
\alias{Cl_test_t}
\alias{Cl_test_t.lm}
\alias{Cl_test_t.default}
\title{Student's t-Test for complex hypothesis for linear models.}
\usage{
Cl_test_t(
  object,
  hypotheses,
  alternatives,
  alpha = 0.05,
  vcov = NULL,
  df = NULL,
  decision.criteria = "both",
  merge. = NULL
)

\method{Cl_test_t}{lm}(
  object,
  hypotheses,
  alternatives,
  alpha = 0.05,
  vcov = NULL,
  decision.criteria = "both",
  merge. = NULL
)

\method{Cl_test_t}{default}(
  object,
  hypotheses,
  alternatives,
  alpha = 0.05,
  vcov,
  df,
  decision.criteria = "both",
  merge. = NULL
)
}
\arguments{
\item{object}{A fitted "lm" object or named vector of coefficients.}

\item{hypotheses}{Vector of null hypothesis.}

\item{alternatives}{Vector of alternative hypothesis, must be one of "not.equal", "greater" or "less".}

\item{alpha}{Float. Statistical significance, by default = 0.05.}

\item{vcov}{Covariance matrix. When supplied with "lm" object, it replaces the vcov matrix from model.}

\item{df}{Number of observation in model diminished by number of exogenous variables with constant term.}

\item{decision.criteria}{Character. Which decision criterion should be displayed when printing result, either "p-value", "critical.region" or "both" (default).}

\item{merge.}{Logical. When FALSE is supplied, identical elements of hypothesis will not be merged.}
}
\value{
A list with class "Clhtest" containing the following components:
}
\description{
Performs multivariate Student's t-Test on linear model.
}
\examples{
x <- runif(100, min = 0, max = 100)
z <- runif(100, min = 0, max = 100)
q <- runif(100, min = 0, max = 100)
e <- rnorm(100, mean = 0, sd = 1)
y <- 4*x + 5*z + 6*q + 1 + e
data <- data.frame(x, z, q, y)
model <- lm(y ~ x + z + q, data)
Cl_test_t(model, c("x=4", "z=5", "q=6"), c("not.equal", "less", "greater"))
}
\seealso{
Other CLH: 
\code{\link{ComplexLinearHypothesis}()}
}
\concept{CLH}
