% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ComplexLinearHypothesis.R
\name{ComplexLinearHypothesis}
\alias{ComplexLinearHypothesis}
\alias{CLH}
\alias{ComplexLinearHypothesis.mvt}
\title{Test Complex Linear Hypothesis}
\usage{
ComplexLinearHypothesis(
  coefficients,
  vcov,
  df,
  hypotheses,
  alternatives,
  alpha = 0.05,
  decision.criteria = "critical.region",
  merge. = NULL,
  family,
  ...
)

CLH(
  coefficients,
  vcov,
  df,
  hypotheses,
  alternatives,
  alpha = 0.05,
  decision.criteria = "critical.region",
  merge. = NULL,
  family,
  ...
)

ComplexLinearHypothesis.mvt(
  coefficients,
  vcov,
  df,
  hypotheses,
  alternatives,
  alpha = 0.05,
  decision.criteria = "critical.region",
  merge. = NULL,
  data.name
)
}
\arguments{
\item{coefficients}{Named vector of coefficients.}

\item{vcov}{Covariance matrix.}

\item{df}{Degrees of freedom.}

\item{hypotheses}{Vector of null hypothesis.}

\item{alternatives}{Vector of alternative hypothesis, must be one of "not.equal", "greater" or "less".}

\item{alpha}{Float. Statistical significance, by default = 0.05.}

\item{decision.criteria}{Character. Which decision criterion should be displayed when printing result, either "p-value" (experimental), "critical.region" (default) or "both."}

\item{merge.}{Logical. When FALSE is supplied, identical elements of hypothesis will not be merged.}

\item{family}{Character. Type of test which should be performed, either "mvt" (multivariate t-Test) or "mvnorm" (multivariate z-Test).}

\item{...}{Arguments to pass down.}
}
\description{
Generic function for testing a complex linear hypothesis, including multivariate
Student's t-Test for generalized linear models (OLS and MM regressions) and
z-Test for MLE parameters.

CLH is an alias for ComplexLinearHypothesis.
}
\seealso{
Other CLH: 
\code{\link{Cl_test_t}()}
}
\concept{CLH}
