% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdnorm.R
\name{mdnorm}
\alias{mdnorm}
\title{The multivariate Normal Distribution}
\usage{
mdnorm(X)
}
\arguments{
\item{X}{Vector of quantiles.}
}
\value{
Numeric value.
}
\description{
Returns density for the m-variate normal distribution for a vector of length 'm'.
}
\examples{
mdnorm(c(1,1))
mdnorm(1)
}
