% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpt.R
\name{mpt}
\alias{mpt}
\title{The multivariate Student t Distribution}
\usage{
mpt(X, v = 1)
}
\arguments{
\item{X}{Vector of quantiles.}

\item{df}{degrees of freedom (>0).}
}
\value{
Numeric value.
}
\description{
Returns CDF for the m-variate t distribution with 'df' degrees of freedom, for a vector of length 'm'.
}
\examples{
mpt(c(1,1), 100)
mpt(1)
}
