% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.test.combo.lm.R
\name{t.test.combo.lm}
\alias{t.test.combo.lm}
\title{Student's t-Test with multiple hypothesis for linear regression model}
\usage{
\method{t}{test.combo.lm}(model, hypotheses, alternatives, alpha = 0.05)
}
\arguments{
\item{model}{A fitted "lm" object.}

\item{hypotheses}{Vector of null hypothesis.}

\item{alternatives}{Vector of alternative hypothesis, must be one of "two.sided", "greater" or "less".}

\item{alpha}{Float. Statistical significance, by default = 0.05.}
}
\value{
An object containg list of hypothesis, value of t-test statistics and corresponding critical regions.
}
\description{
Student's t-Test with multiple hypothesis for linear regression model
}
\examples{
x <- runif(100, min = 0, max = 100)
z <- runif(100, min = 0, max = 100)
q <- runif(100, min = 0, max = 100)
e <- rnorm(100, mean = 0, sd = 1)
y <- 4*x + 5*z + 6*q + 1 + e
data <- data.frame(x, z, q, y)
model <- lm(y ~ x + z + q, data)
MultiTest::t.test.combo.lm(model, c("x=4", "z=5", "q=6"), c("not.equal", "less", "greater"))
}
