% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_combo.df.R
\name{t_test_combo.df}
\alias{t_test_combo.df}
\title{Student's t-Test with multiple hypothesis}
\usage{
\method{t_test_combo}{df}(object, hypotheses, alternatives, alpha = 0.05)
}
\arguments{
\item{object}{A data frame.}

\item{hypotheses}{Vector of null hypothesis.}

\item{alternatives}{Vector of alternative hypothesis, must be one of "two.sided", "greater" or "less".}

\item{alpha}{Float. Statistical significance, by default = 0.05.}
}
\value{
An object containg list of hypothesis, value of t-test statistics and corresponding critical regions.
}
\description{
Student's t-Test with multiple hypothesis
}
\examples{
x <- runif(100, min = 0, max = 100)
z <- runif(100, min = 0, max = 100)
q <- runif(100, min = 0, max = 100)
data <- data.frame(x, z, q)
t_test_combo.df(data, c("x=4", "z=5", "q=6"), c("not.equal", "less", "greater"))
}
